+++ src/v2/view-builder/components/IdentifierFooter.js
@@ -1,7 +1,7 @@
-import { loc, _ } from 'okta';
+import { loc } from 'okta';
 import { BaseFooter } from '../internals';
 import { FORMS as RemediationForms } from '../../ion/RemediationConstants';
-import { getForgotPasswordLink } from '../utils/LinksUtil';
+import { getForgotPasswordLink, getSignUpLink } from '../utils/LinksUtil';
 
 export default BaseFooter.extend({
   links () {
@@ -21,20 +21,7 @@ export default BaseFooter.extend({
       },
     ];
 
-    const signupLink = [];
-    if (this.options.appState.hasRemediationObject(RemediationForms.SELECT_ENROLL_PROFILE)) {
-      const signupLinkData = {
-        'type': 'link',
-        'label': loc('signup', 'login'),
-        'name': 'enroll',
-      };
-      if (_.isFunction(this.options.settings.get('registration.click'))) {
-        signupLinkData.click = this.options.settings.get('registration.click');
-      } else {
-        signupLinkData.actionPath = RemediationForms.SELECT_ENROLL_PROFILE;
-      }
-      signupLink.push(signupLinkData);
-    }
+    const signUpLink = getSignUpLink(this.options.appState, this.options.settings);
 
     const forgotPasswordLink = getForgotPasswordLink(this.options.appState, this.options.settings);
 
@@ -60,7 +47,7 @@ export default BaseFooter.extend({
 
     return forgotPasswordLink
       .concat(unlockAccountLink)
-      .concat(signupLink)
+      .concat(signUpLink)
       .concat(helpLink)
       .concat(customHelpLinks);
   }
